[fix/experience-null-check] Null-check on response of experiences endpoint.
Up to now, there was an assumption that response of /experiences endpoint always contains value of 'experiencePayloads' property, even if there are no experiences.
It is the most probable reason of crash reported by client.
Now, experiences connector is prepared for receiving null value and treats it as empty array.